Re-structure the classes
