+++ urdfpy/urdf/manipulation.py
@@ -10,727 +10,9 @@ import open3d as o3d
 import six
 
 from urdfpy.urdf.base import URDFType
-from urdfpy.urdf.joint import (
-    JointCalibration,
-    JointDynamics,
-    JointLimit,
-    JointMimic,
-    SafetyController
-)
-from urdfpy.urdf.link import Collision, Material, Visual
-from urdfpy.urdf.transmission import TransmissionJoint, Actuator
-from urdfpy.utils import parse_origin, unparse_origin, configure_origin
-
-
-class Transmission(URDFType):
-    """An element that describes the relationship between an actuator and a
-    joint.
-
-    Parameters
-    ----------
-    name : str
-        The name of this transmission.
-    trans_type : str
-        The type of this transmission.
-    joints : list of :class:`.TransmissionJoint`
-        The joints connected to this transmission.
-    actuators : list of :class:`.Actuator`
-        The actuators connected to this transmission.
-    """
-    _ATTRIBS = {
-        'name': (str, True),
-    }
-    _ELEMENTS = {
-        'joints': (TransmissionJoint, True, True),
-        'actuators': (Actuator, True, True),
-    }
-    _TAG = 'transmission'
-
-    def __init__(self, name, trans_type, joints=None, actuators=None):
-        self.name = name
-        self.trans_type = trans_type
-        self.joints = joints
-        self.actuators = actuators
-
-    @property
-    def name(self):
-        """str : The name of this transmission.
-        """
-        return self._name
-
-    @name.setter
-    def name(self, value):
-        self._name = str(value)
-
-    @property
-    def trans_type(self):
-        """str : The type of this transmission.
-        """
-        return self._trans_type
-
-    @trans_type.setter
-    def trans_type(self, value):
-        self._trans_type = str(value)
-
-    @property
-    def joints(self):
-        """:class:`.TransmissionJoint` : The joints the transmission is
-        connected to.
-        """
-        return self._joints
-
-    @joints.setter
-    def joints(self, value):
-        if value is None:
-            value = []
-        else:
-            value = list(value)
-            for v in value:
-                if not isinstance(v, TransmissionJoint):
-                    raise TypeError(
-                        'Joints expects a list of TransmissionJoint'
-                    )
-        self._joints = value
-
-    @property
-    def actuators(self):
-        """:class:`.Actuator` : The actuators the transmission is connected to.
-        """
-        return self._actuators
-
-    @actuators.setter
-    def actuators(self, value):
-        if value is None:
-            value = []
-        else:
-            value = list(value)
-            for v in value:
-                if not isinstance(v, Actuator):
-                    raise TypeError(
-                        'Actuators expects a list of Actuator'
-                    )
-        self._actuators = value
-
-    @classmethod
-    def _from_xml(cls, node, path):
-        kwargs = cls._parse(node, path)
-        kwargs['trans_type'] = node.find('type').text
-        return Transmission(**kwargs)
-
-    def _to_xml(self, parent, path):
-        node = self._unparse(path)
-        ttype = ET.Element('type')
-        ttype.text = self.trans_type
-        node.append(ttype)
-        return node
-
-    def copy(self, prefix='', scale=None):
-        """Create a deep copy with the prefix applied to all names.
-
-        Parameters
-        ----------
-        prefix : str
-            A prefix to apply to all names.
-
-        Returns
-        -------
-        :class:`.Transmission`
-            A deep copy.
-        """
-        return Transmission(
-            name='{}{}'.format(prefix, self.name),
-            trans_type=self.trans_type,
-            joints=[j.copy(prefix) for j in self.joints],
-            actuators=[a.copy(prefix) for a in self.actuators],
-        )
-
-
-class Joint(URDFType):
-    """A connection between two links.
-
-    There are several types of joints, including:
-
-    - ``fixed`` - a joint that cannot move.
-    - ``prismatic`` - a joint that slides along the joint axis.
-    - ``revolute`` - a hinge joint that rotates about the axis with a limited
-      range of motion.
-    - ``continuous`` - a hinge joint that rotates about the axis with an
-      unlimited range of motion.
-    - ``planar`` - a joint that moves in the plane orthogonal to the axis.
-    - ``floating`` - a joint that can move in 6DoF.
-
-    Parameters
-    ----------
-    name : str
-        The name of this joint.
-    parent : str
-        The name of the parent link of this joint.
-    child : str
-        The name of the child link of this joint.
-    joint_type : str
-        The type of the joint. Must be one of :obj:`.Joint.TYPES`.
-    axis : (3,) float, optional
-        The axis of the joint specified in joint frame. Defaults to
-        ``[1,0,0]``.
-    origin : (4,4) float, optional
-        The pose of the child link with respect to the parent link's frame.
-        The joint frame is defined to be coincident with the child link's
-        frame, so this is also the pose of the joint frame with respect to
-        the parent link's frame.
-    limit : :class:`.JointLimit`, optional
-        Limit for the joint. Only required for revolute and prismatic
-        joints.
-    dynamics : :class:`.JointDynamics`, optional
-        Dynamics for the joint.
-    safety_controller : :class`.SafetyController`, optional
-        The safety controller for this joint.
-    calibration : :class:`.JointCalibration`, optional
-        Calibration information for the joint.
-    mimic : :class:`JointMimic`, optional
-        Joint mimicry information.
-    """
-    TYPES = ['fixed', 'prismatic', 'revolute',
-             'continuous', 'floating', 'planar']
-    _ATTRIBS = {
-        'name': (str, True),
-    }
-    _ELEMENTS = {
-        'dynamics': (JointDynamics, False, False),
-        'limit': (JointLimit, False, False),
-        'mimic': (JointMimic, False, False),
-        'safety_controller': (SafetyController, False, False),
-        'calibration': (JointCalibration, False, False),
-    }
-    _TAG = 'joint'
-
-    def __init__(self, name, joint_type, parent, child, axis=None, origin=None,
-                 limit=None, dynamics=None, safety_controller=None,
-                 calibration=None, mimic=None):
-        self.name = name
-        self.parent = parent
-        self.child = child
-        self.joint_type = joint_type
-        self.axis = axis
-        self.origin = origin
-        self.limit = limit
-        self.dynamics = dynamics
-        self.safety_controller = safety_controller
-        self.calibration = calibration
-        self.mimic = mimic
-
-    @property
-    def name(self):
-        """str : Name for this joint.
-        """
-        return self._name
-
-    @name.setter
-    def name(self, value):
-        self._name = str(value)
-
-    @property
-    def joint_type(self):
-        """str : The type of this joint.
-        """
-        return self._joint_type
-
-    @joint_type.setter
-    def joint_type(self, value):
-        value = str(value)
-        if value not in Joint.TYPES:
-            raise ValueError('Unsupported joint type {}'.format(value))
-        self._joint_type = value
-
-    @property
-    def parent(self):
-        """str : The name of the parent link.
-        """
-        return self._parent
-
-    @parent.setter
-    def parent(self, value):
-        self._parent = str(value)
-
-    @property
-    def child(self):
-        """str : The name of the child link.
-        """
-        return self._child
-
-    @child.setter
-    def child(self, value):
-        self._child = str(value)
-
-    @property
-    def axis(self):
-        """(3,) float : The joint axis in the joint frame.
-        """
-        return self._axis
-
-    @axis.setter
-    def axis(self, value):
-        if value is None:
-            value = np.array([1.0, 0.0, 0.0], dtype=np.float64)
-        else:
-            value = np.asanyarray(value, dtype=np.float64)
-            if value.shape != (3,):
-                raise ValueError('Invalid shape for axis, should be (3,)')
-            value = value / np.linalg.norm(value)
-        self._axis = value
-
-    @property
-    def origin(self):
-        """(4,4) float : The pose of child and joint frames relative to the
-        parent link's frame.
-        """
-        return self._origin
-
-    @origin.setter
-    def origin(self, value):
-        self._origin = configure_origin(value)
-
-    @property
-    def limit(self):
-        """:class:`.JointLimit` : The limits for this joint.
-        """
-        return self._limit
-
-    @limit.setter
-    def limit(self, value):
-        if value is None:
-            if self.joint_type in ['prismatic', 'revolute']:
-                raise ValueError('Require joint limit for prismatic and '
-                                 'revolute joints')
-        elif not isinstance(value, JointLimit):
-            raise TypeError('Expected JointLimit type')
-        self._limit = value
-
-    @property
-    def dynamics(self):
-        """:class:`.JointDynamics` : The dynamics for this joint.
-        """
-        return self._dynamics
-
-    @dynamics.setter
-    def dynamics(self, value):
-        if value is not None:
-            if not isinstance(value, JointDynamics):
-                raise TypeError('Expected JointDynamics type')
-        self._dynamics = value
-
-    @property
-    def safety_controller(self):
-        """:class:`.SafetyController` : The safety controller for this joint.
-        """
-        return self._safety_controller
-
-    @safety_controller.setter
-    def safety_controller(self, value):
-        if value is not None:
-            if not isinstance(value, SafetyController):
-                raise TypeError('Expected SafetyController type')
-        self._safety_controller = value
-
-    @property
-    def calibration(self):
-        """:class:`.JointCalibration` : The calibration for this joint.
-        """
-        return self._calibration
-
-    @calibration.setter
-    def calibration(self, value):
-        if value is not None:
-            if not isinstance(value, JointCalibration):
-                raise TypeError('Expected JointCalibration type')
-        self._calibration = value
-
-    @property
-    def mimic(self):
-        """:class:`.JointMimic` : The mimic for this joint.
-        """
-        return self._mimic
-
-    @mimic.setter
-    def mimic(self, value):
-        if value is not None:
-            if not isinstance(value, JointMimic):
-                raise TypeError('Expected JointMimic type')
-        self._mimic = value
-
-    def is_valid(self, cfg):
-        """Check if the provided configuration value is valid for this joint.
-
-        Parameters
-        ----------
-        cfg : float, (2,) float, (6,) float, or (4,4) float
-            The configuration of the joint.
-
-        Returns
-        -------
-        is_valid : bool
-            True if the configuration is valid, and False otherwise.
-        """
-        if self.joint_type not in ['fixed', 'revolute']:
-            return True
-        if self.joint_limit is None:
-            return True
-        cfg = float(cfg)
-        lower = -np.infty
-        upper = np.infty
-        if self.limit.lower is not None:
-            lower = self.limit.lower
-        if self.limit.upper is not None:
-            upper = self.limit.upper
-        return (cfg >= lower and cfg <= upper)
-
-    def get_child_pose(self, cfg=None):
-        """Computes the child pose relative to a parent pose for a given
-        configuration value.
-
-        Parameters
-        ----------
-        cfg : float, (2,) float, (6,) float, or (4,4) float
-            The configuration values for this joint. They are interpreted
-            based on the joint type as follows:
-
-            - ``fixed`` - not used.
-            - ``prismatic`` - a translation along the axis in meters.
-            - ``revolute`` - a rotation about the axis in radians.
-            - ``continuous`` - a rotation about the axis in radians.
-            - ``planar`` - the x and y translation values in the plane.
-            - ``floating`` - the xyz values followed by the rpy values,
-              or a (4,4) matrix.
-
-            If ``cfg`` is ``None``, then this just returns the joint pose.
-
-        Returns
-        -------
-        pose : (4,4) float
-            The pose of the child relative to the parent.
-        """
-        if cfg is None:
-            return self.origin
-        elif self.joint_type == 'fixed':
-            return self.origin
-        elif self.joint_type in ['revolute', 'continuous']:
-            if cfg is None:
-                angle = 0.0
-            else:
-                angle = float(cfg)
-            R = np.eye(4)
-            R[:3, :3] = o3d.geometry.get_rotation_matrix_from_axis_angle(angle * self.axis)
-            return self.origin.dot(R)
-        elif self.joint_type == 'prismatic':
-            if cfg is None:
-                cfg = 0.0
-            else:
-                cfg = float(cfg)
-            translation = np.eye(4, dtype=np.float64)
-            translation[:3,3] = self.axis * cfg
-            return self.origin.dot(translation)
-        elif self.joint_type == 'planar':
-            if cfg is None:
-                cfg = np.zeros(2, dtype=np.float64)
-            else:
-                cfg = np.asanyarray(cfg, dtype=np.float64)
-            if cfg.shape != (2,):
-                raise ValueError(
-                    '(2,) float configuration required for planar joints'
-                )
-            translation = np.eye(4, dtype=np.float64)
-            translation[:3,3] = self.origin[:3,:2].dot(cfg)
-            return self.origin.dot(translation)
-        elif self.joint_type == 'floating':
-            if cfg is None:
-                cfg = np.zeros(6, dtype=np.float64)
-            else:
-                cfg = configure_origin(cfg)
-            if cfg is None:
-                raise ValueError('Invalid configuration for floating joint')
-            return self.origin.dot(cfg)
-        else:
-            raise ValueError('Invalid configuration')
-
-    def get_child_poses(self, cfg, n_cfgs):
-        """Computes the child pose relative to a parent pose for a given set of 
-        configuration values.
-
-        Parameters
-        ----------
-        cfg : (n,) float or None
-            The configuration values for this joint. They are interpreted
-            based on the joint type as follows:
-
-            - ``fixed`` - not used.
-            - ``prismatic`` - a translation along the axis in meters.
-            - ``revolute`` - a rotation about the axis in radians.
-            - ``continuous`` - a rotation about the axis in radians.
-            - ``planar`` - Not implemented.
-            - ``floating`` - Not implemented.
-
-            If ``cfg`` is ``None``, then this just returns the joint pose.
-
-        Returns
-        -------
-        poses : (n,4,4) float
-            The poses of the child relative to the parent.
-        """
-        if cfg is None:
-            return np.tile(self.origin, (n_cfgs, 1, 1))
-        elif self.joint_type == 'fixed':
-            return np.tile(self.origin, (n_cfgs, 1, 1))
-        elif self.joint_type in ['revolute', 'continuous']:
-            if cfg is None:
-                cfg = np.zeros(n_cfgs)
-            return np.matmul(self.origin, self._rotation_matrices(cfg, self.axis))
-        elif self.joint_type == 'prismatic':
-            if cfg is None:
-                cfg = np.zeros(n_cfgs)
-            translation = np.tile(np.eye(4), (n_cfgs, 1, 1))
-            translation[:,:3,3] = self.axis * cfg[:,np.newaxis]
-            return np.matmul(self.origin, translation)
-        elif self.joint_type == 'planar':
-            raise NotImplementedError()
-        elif self.joint_type == 'floating':
-            raise NotImplementedError()
-        else:
-            raise ValueError('Invalid configuration')
-
-    @classmethod
-    def _from_xml(cls, node, path):
-        kwargs = cls._parse(node, path)
-        kwargs['joint_type'] = str(node.attrib['type'])
-        kwargs['parent'] = node.find('parent').attrib['link']
-        kwargs['child'] = node.find('child').attrib['link']
-        axis = node.find('axis')
-        if axis is not None:
-            axis = np.fromstring(axis.attrib['xyz'], sep=' ')
-        kwargs['axis'] = axis
-        kwargs['origin'] = parse_origin(node)
-        return Joint(**kwargs)
-
-    def _to_xml(self, parent, path):
-        node = self._unparse(path)
-        parent = ET.Element('parent')
-        parent.attrib['link'] = self.parent
-        node.append(parent)
-        child = ET.Element('child')
-        child.attrib['link'] = self.child
-        node.append(child)
-        if self.axis is not None:
-            axis = ET.Element('axis')
-            axis.attrib['xyz'] = np.array2string(self.axis)[1:-1]
-            node.append(axis)
-        node.append(unparse_origin(self.origin))
-        node.attrib['type'] = self.joint_type
-        return node
-
-    def _rotation_matrices(self, angles, axis):
-        """Compute rotation matrices from angle/axis representations.
-
-        Parameters
-        ----------
-        angles : (n,) float
-            The angles.
-        axis : (3,) float
-            The axis.
-
-        Returns
-        -------
-        rots : (n,4,4)
-            The rotation matrices
-        """
-        axis = axis / np.linalg.norm(axis)
-        sina = np.sin(angles)
-        cosa = np.cos(angles)
-        M = np.tile(np.eye(4), (len(angles), 1, 1))
-        M[:,0,0] = cosa
-        M[:,1,1] = cosa
-        M[:,2,2] = cosa
-        M[:,:3,:3] += (
-            np.tile(np.outer(axis, axis), (len(angles), 1, 1)) *
-            (1.0 - cosa)[:, np.newaxis, np.newaxis]
-        )
-        M[:,:3,:3] += np.tile(np.array([
-            [0.0, -axis[2], axis[1]],
-            [axis[2], 0.0, -axis[0]],
-            [-axis[1], axis[0], 0.0]]
-        ), (len(angles), 1, 1)) * sina[:, np.newaxis, np.newaxis]
-        return M
-
-    def copy(self, prefix='', scale=None):
-        """Create a deep copy of the joint with the prefix applied to all names.
-
-        Parameters
-        ----------
-        prefix : str
-            A prefix to apply to all joint and link names.
-
-        Returns
-        -------
-        :class:`.Joint`
-            A deep copy of the joint.
-        """
-        origin = self.origin.copy()
-        if scale is not None:
-            if not isinstance(scale, (list, np.ndarray)):
-                scale = np.repeat(scale, 3)
-            origin[:3,3] *= scale
-        cpy = Joint(
-            name='{}{}'.format(prefix, self.name),
-            joint_type=self.joint_type,
-            parent='{}{}'.format(prefix, self.parent),
-            child='{}{}'.format(prefix, self.child),
-            axis=self.axis.copy(),
-            origin=origin,
-            limit=(self.limit.copy(prefix, scale) if self.limit else None),
-            dynamics=(self.dynamics.copy(prefix,scale) if self.dynamics else None),
-            safety_controller=(self.safety_controller.copy(prefix, scale) if
-                               self.safety_controller else None),
-            calibration=(self.calibration.copy(prefix, scale) if self.calibration else None),
-            mimic=(self.mimic.copy(prefix=prefix, scale=scale) if self.mimic else None)
-        )
-        return cpy
-
-
-class Link(URDFType):
-    """A link of a rigid object.
-
-    Parameters
-    ----------
-    name : str
-        The name of the link.
-    visuals : list of :class:`.Visual`, optional
-        The visual properties of the link.
-    collsions : list of :class:`.Collision`, optional
-        The collision properties of the link.
-    """
-
-    _ATTRIBS = {
-        'name': (str, True),
-    }
-    _ELEMENTS = {
-        'visuals': (Visual, False, True),
-        'collisions': (Collision, False, True),
-    }
-    _TAG = 'link'
-
-    def __init__(self, name, visuals, collisions):
-        self.name = name
-        self.visuals = visuals
-        self.collisions = collisions
-
-        self._collision_mesh = None
-
-    @property
-    def name(self):
-        """str : The name of this link.
-        """
-        return self._name
-
-    @name.setter
-    def name(self, value):
-        self._name = str(value)
-
-    @property
-    def visuals(self):
-        """list of :class:`.Visual` : The visual properties of this link.
-        """
-        return self._visuals
-
-    @visuals.setter
-    def visuals(self, value):
-        if value is None:
-            value = []
-        else:
-            value = list(value)
-            for v in value:
-                if not isinstance(v, Visual):
-                    raise ValueError('Expected list of Visual objects')
-        self._visuals = value
-
-    @property
-    def collisions(self):
-        """list of :class:`.Collision` : The collision properties of this link.
-        """
-        return self._collisions
-
-    @collisions.setter
-    def collisions(self, value):
-        if value is None:
-            value = []
-        else:
-            value = list(value)
-            for v in value:
-                if not isinstance(v, Collision):
-                    raise ValueError('Expected list of Collision objects')
-        self._collisions = value
-
-    @property
-    def collision_mesh(self):
-        """:class:`~trimesh.base.Trimesh` : A single collision mesh for
-        the link, specified in the link frame, or None if there isn't one.
-        """
-        if len(self.collisions) == 0:
-            return None
-        if self._collision_mesh is None:
-            meshesList = []
-            # list of meshes from ALL collision objects
-            for c in self.collisions:
-                m = c.geometry.meshes
-                # NOTE: m here is a o3d.geometry.TriangleMesh
-                m = copy.deepcopy(m)
-                pose = c.origin
-                if c.geometry.mesh is not None:
-                    # c.geometry.mesh  is  the  Mesh  object,  an
-                    # instance of the derived class from URDFType
-                    if c.geometry.mesh.scale is not None:
-                        S = np.eye(4)
-                        S[:3,:3] = np.diag(c.geometry.mesh.scale)
-                        pose = pose.dot(S)
-                m.transform(pose)
-                meshesList.append(m)
-            if len(meshesList) == 0:
-                return None
-            self._collision_mesh = meshesList[0]
-            for otherMesh in meshesList[1:]:
-                self._collision_mesh = self._collision_mesh + otherMesh
-        return self._collision_mesh
-
-    def copy(self, prefix='', scale=None, collision_only=False):
-        """Create a deep copy of the link.
-
-        Parameters
-        ----------
-        prefix : str
-            A prefix to apply to all joint and link names.
-
-        Returns
-        -------
-        link : :class:`.Link`
-            A deep copy of the Link.
-        """
-        cm = self._collision_mesh
-        if scale is not None and self.collision_mesh is not None:
-            sm = np.eye(4)
-            if not isinstance(scale, (list, np.ndarray)):
-                scale = np.repeat(scale, 3)
-            sm[:3,:3] = np.diag(scale)
-            cm = copy.deepcopy(self.collision_mesh)
-            cm.transform(sm)
-
-        visuals = None
-        if not collision_only:
-            visuals=[v.copy(prefix=prefix, scale=scale) for v in self.visuals]
-
-        cpy = Link(
-            name='{}{}'.format(prefix, self.name),
-            visuals=visuals,
-            collisions=[v.copy(prefix=prefix, scale=scale) for v in self.collisions],
-        )
-        cpy._collision_mesh = cm
-        return cpy
-
+from urdfpy.urdf.joint import Joint
+from urdfpy.urdf.transmission import Transmission
+from urdfpy.urdf.link import Material, Link
 
 class URDF(URDFType):
     """The top-level URDF specification.
